更改了六边形格子寻找周围格子的算法，放弃了碰撞检测。 格子从被动接受攻击改为主动攻击

+++ 2DGame/Assets/scripts/BatteryData.cs
@@ -8,7 +8,7 @@ public class BatteryData
     public GameObject batteryPrefab;
     public int cost;
     public BatteryType type;
-    public float cureDistance;
+    public int cureDistance;
     public float cureVal;
     public float totalHP;
     public float HP; //当前血量
